microservices/review/pyproject.toml
[tool.poetry]
name = "review"
version = "0.1.0"
description = ""
authors = ["Clariteia Devs <devs@clariteia.com>"]
packages = [{ include = "src" }]

[tool.poetry.dependencies]
python = "^3.9"
minos-microservice-common = "^0.1.9"
minos-microservice-networks = ">=0.0.12<0.1"
minos-microservice-saga = ">=0.0.9<0.1"
minos-microservice-cqrs = ">=0.0.1<0.1"
typer = "^0.3.2"
SQLAlchemy = "^1.4.22"

[tool.poetry.dev-dependencies]
pytest = "^6.2.4"
black = "^19.10b"
isort = "^5.8.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
microservice = "src.cli:main"